Delete night_writer_test file since nigh_writer is now a runner and no longer a class
